Dockerfile fix, pipeline generates image, configuration db and mqtt
Dockerfile is fixed. There is new pipeline script that can be in the future used for building and testing the project, but right now it just automatically generates Docker image for this project. When user wants to build docker image for this project, he/she no longer should do it manually via Dockerfile, because it is OS dependand (on Windows, all newlines in .sh and .py files are replaced, so the image fails after run). From now on, user should use the automatically built image that can be found at GitLab container registry. See README for more details.